refactor(core): extract incidents page into 13 files #159
